modificacion en botton de google correcion de bugs
